go-tour: convert from HTML to present format
This makes it easier for us to maintain the tour, and puts the tour
sample programs in individual go files that are compilable and
testable.

This also removes the code that shows content specific to the appengine
or local versions, erring on the side of inclusivity. I intend to add
some more appengine-specific stuff in a followup CL.

R=campoy, bradfitz, campoy, rsc
CC=golang-dev
https://codereview.appspot.com/6896044